# openapi.yaml
openapi: 3.0.3
info:
  title: Clearbid
  version: 0.1.0
  description: >-
    Cost-per-action advertising exchange that settles media spend against mutually
    verified conversions. Advertisers buy target actions, publishers receive payouts
    backed by an append-only conversion ledger, and invalid traffic is excluded
    before settlement rather than clawed back afterwards.
  x-product-slug: ai-advertising-clearbid
  x-source: ai-in-advertizing/Ubex-Whitepaper-en/
servers:
  - url: https://api.example.com
...
paths:
  /v1/campaigns:
    get:
      operationId: listCampaigns
      tags: [Campaigns]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, learning, active, paused, ended]
      responses:
        "200":
          description: Campaigns page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeCampaign"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createCampaign
      tags: [Campaigns]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignCreate"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeCampaign"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /v1/campaigns/{campaignId}:
    parameters:
      - $ref: "#/components/parameters/CampaignId"
    get:
      operationId: getCampaign
      tags: [Campaigns]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeCampaign"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/conversions:
    get:
      operationId: listConversions
      tags: [Conversions]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: campaignId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [attested, confirmed, excluded, disputed, settled]
      responses:
        "200":
          description: Conversion records page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeConversionRecord"
    post:
      operationId: attestConversion
      tags: [Conversions]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversionAttestation"
      responses:
        "201":
          description: Conversion attested and written to the ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeConversionRecord"
        "409":
          description: Attribution token already redeemed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    CampaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: double
        currency:
          type: string
          example: USD
    Campaign:
      type: object
      required: [id, advertiserId, name, status, pricingModel]
      properties:
        id:
          type: string
        advertiserId:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [draft, learning, active, paused, ended]
        pricingModel:
          type: string
          enum: [cost_per_action, cost_per_click, cost_per_mille]
        maxPricePerAction:
          $ref: "#/components/schemas/Money"
        budget:
          $ref: "#/components/schemas/Money"
        brandSafetyCategories:
          type: array
          items:
            type: string
        consentBasis:
          type: string
          enum: [consent, legitimate_interest]
    CampaignCreate:
      type: object
      required: [name, pricingModel, budget]
      properties:
        name:
          type: string
        pricingModel:
          type: string
          enum: [cost_per_action, cost_per_click, cost_per_mille]
        maxPricePerAction:
          $ref: "#/components/schemas/Money"
        budget:
          $ref: "#/components/schemas/Money"
        brandSafetyCategories:
          type: array
          items:
            type: string
...
    ConversionAttestation:
      type: object
      required: [attributionToken, actionType, occurredAt]
      properties:
        attributionToken:
          type: string
        actionType:
          type: string
          enum: [registration, form_completion, callback_request, purchase]
        occurredAt:
          type: string
          format: date-time
        externalReference:
          type: string
          description: Advertiser system-of-record identifier, no personal data
    ConversionRecord:
      type: object
      required: [id, campaignId, slotId, status, actionType]
      properties:
        id:
          type: string
        campaignId:
          type: string
        slotId:
          type: string
        actionType:
          type: string
        status:
          type: string
          enum: [attested, confirmed, excluded, disputed, settled]
        exclusionReason:
          type: string
        payableAmount:
          $ref: "#/components/schemas/Money"
        ledgerSequence:
          type: integer
...
    DataEnvelopeCampaign:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Campaign"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    DataEnvelopeConversionRecord:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/ConversionRecord"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ListEnvelopeCampaign:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Campaign"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ListEnvelopeConversionRecord:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ConversionRecord"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...