# platform/api-server/src/domains/campaigns/handlers/campaign.handlers.ts
/**
 * Handlers for Campaign
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: campaigns
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { CampaignsDomainModule } from "../dependencies/campaigns-ddd.dependencies.js";
import type { CreateCampaignInput, GetCampaignInput, ListCampaignsInput } from "@clearbid/services/campaigns";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/campaigns
 * Operation: listCampaigns
 */
export async function listCampaigns(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], status?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: CampaignsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListCampaignsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.campaigns.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/campaigns
 * Operation: createCampaign
 */
export async function createCampaign(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: CampaignsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: CreateCampaignInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.campaigns.create.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v1/campaigns/{campaignId}
 * Operation: getCampaign
 */
export async function getCampaign(
  request: FastifyRequest<{ Params: { campaignId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: CampaignsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetCampaignInput = {
    campaignId: request.params.campaignId,
  };
  const result = await deps.useCases.campaigns.get.execute(input);
  return reply.code(200).send(result);
}