# platform/api-server/src/domains/conversions/handlers/conversion.handlers.ts
/**
 * Handlers for Conversion
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: conversions
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { ConversionsDomainModule } from "../dependencies/conversions-ddd.dependencies.js";
import type { AttestConversionInput, ListConversionsInput } from "@clearbid/services/conversions";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/conversions
 * Operation: listConversions
 */
export async function listConversions(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], status?: string | string[], campaignId?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: ConversionsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListConversionsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.conversions.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/conversions
 * Operation: attestConversion
 */
export async function attestConversion(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: ConversionsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: AttestConversionInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.conversions.get.execute(input);
  return reply.code(201).send(result);
}